requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pwdlib[argon2]
pyjwt
python-multipart
python-dotenv